+++ CommandCenter/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // 有关程序集的一般信息由以下
 // 控制。更改这些特性值可修改
 // 与程序集关联的信息。
-[assembly: AssemblyTitle("CommandCenter")]
-[assembly: AssemblyDescription("命令中心上位机软件")]
+[assembly: AssemblyTitle("光阑视界 IrisVision")]
+[assembly: AssemblyDescription("光阑视界 IrisVision 上位机软件")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("CommandCenter")]
+[assembly: AssemblyProduct("IrisVision")]
 [assembly: AssemblyCopyright("Copyright ©  2026")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]